+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 - fixed phone number validation on block-based Checkout page
 
+### Improved
+- optimized loading of the plugin settings page so as not to slow down the website
+
 ## [1.20.4]
 ### Fixed
 - fixed so that the shipping method class would be loaded earlier